Updated timescheme to interface
