+++ docs/api/solver.rst
@@ -3,5 +3,8 @@
 Solver interface
 ================
 
-.. doxygenfile:: solver.h
+.. doxygenfile:: solver.hpp
+    :project: SPECFEM KOKKOS IMPLEMENTATION
+
+.. doxygenfile:: time_marching.hpp
     :project: SPECFEM KOKKOS IMPLEMENTATION
