+++ include/timescheme/interface.hpp
@@ -1,0 +1,7 @@
+#ifndef _TIMESCHEME_INTERFACE_HPP
+#define _TIMESCHEME_INTERFACE_HPP
+
+#include "newmark.hpp"
+#include "timescheme.hpp"
+
+#endif
