+++ tests/unittests/seismogram/seismogram_tests.cpp
@@ -1,16 +1,16 @@
-#include "../../../include/compute.h"
-#include "../../../include/domain.h"
-#include "../../../include/material.h"
-#include "../../../include/mesh.h"
-#include "../../../include/parameter_parser.h"
-#include "../../../include/quadrature.h"
-#include "../../../include/read_sources.h"
-#include "../../../include/solver/interface.hpp"
-#include "../../../include/timescheme.h"
-#include "../../../include/utils.h"
 #include "../Kokkos_Environment.hpp"
 #include "../MPI_environment.hpp"
 #include "../utilities/include/compare_array.h"
+#include "compute.h"
+#include "domain.h"
+#include "material.h"
+#include "mesh.h"
+#include "parameter_parser.h"
+#include "quadrature.h"
+#include "read_sources.h"
+#include "solver/interface.hpp"
+#include "timescheme/interface.hpp"
+#include "utils.h"
 #include "yaml-cpp/yaml.h"
 
 // ----- Parse test config ------------- //
